tests: create reusable ipc-client helper
